collection progress bar, ranking 페이지 구현

+++ main/models.py
@@ -14,8 +14,8 @@ class User(AbstractUser):
 class Landmark(models.Model):
     lanmark_id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=50)
-    x = models.IntegerField()
-    y = models.IntegerField()
+    x = models.FloatField()
+    y = models.FloatField()
     area = models.CharField(max_length=50)
 
     class Meta:
